Rearranging prevoius xfail tests since they use default.qubit specifically; fix license
